+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 - add button with mass change options menu to hero property tabs
 - add individual row options menu to hero skills, army, equipment, and inventory
 - add option to refill hero movement points
+- add info text to inventory artifacts with hero main stats bonuses
 - add info text to spells on missing spellbook